Renamed BuildDefs to PlatformDefs
