+++ cmake/project_definitions.cmake
@@ -51,7 +51,7 @@ set(TARGETNAME_INCLUDES_SDK sdk_includes)
 set(TARGETNAME_INCLUDES_FILESYSTEM filesystem_includes)
 set(TARGETNAME_INCLUDES_ENGINE engine_includes)
 
-set(TARGETNAME_LIB_BUILDDEFS builddefs)
+set(TARGETNAME_LIB_PLATFORMDEFS platformdefs)
 set(TARGETNAME_LIB_COMMONUTILS commonutils)
 set(TARGETNAME_LIB_FILESYSTEM_STDIO filesystem_stdio)
 set(TARGETNAME_LIB_FILESYSTEM_TESTS filesystem_test_caseinsensitive)
