+++ game/game_libs/ui/CMakeLists.txt
@@ -224,7 +224,7 @@ target_link_libraries(${TARGETNAME_LIB_UI} PRIVATE
 	${TARGETNAME_LIB_MINIUTL}
 	${TARGETNAME_INCLUDES_SDK}
 	${TARGETNAME_LIB_PLATFORMLIB}
-	${TARGETNAME_LIB_BUILDDEFS}
+	${TARGETNAME_LIB_PLATFORMDEFS}
 	${TARGETNAME_LIB_COMMONUTILS}
 
 	# Windows
