+++ libraries/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(platformlib)
 add_subdirectory(thirdparty)
-add_subdirectory(builddefs)
+add_subdirectory(platformdefs)
 add_subdirectory(commonutils)
 add_subdirectory(crclib)
