+++ libraries/commonutils/CMakeLists.txt
@@ -22,7 +22,7 @@ target_include_directories(${TARGETNAME_LIB_COMMONUTILS}
 target_link_libraries(${TARGETNAME_LIB_COMMONUTILS}
 	PUBLIC
 	${TARGETNAME_LIB_PLATFORMLIB}
-	${TARGETNAME_LIB_BUILDDEFS}
+	${TARGETNAME_LIB_PLATFORMDEFS}
 
 	PRIVATE
 	$<$<STREQUAL:"${TARGET_PLATFORM}","LINUX">:m>
