+++ libraries/platformdefs/include/PlatformDefs/decorators.h
