+++ xash3d_engine/engine/client/avi/avi_win.c
@@ -14,7 +14,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "BuildDefs/build.h"
+#include "PlatformDefs/build.h"
 #if XASH_WIN32
 #include "common.h"
 #include "client.h"
