+++ xash3d_engine/engine/client/cl_gameui.c
@@ -20,7 +20,7 @@ GNU General Public License for more details.
 #include "input.h"
 #include "server.h"  // !!svgame.hInstance
 #include "vid_common.h"
-#include "BuildDefs/libnames.h"
+#include "PlatformDefs/libnames.h"
 #include <limits.h>
 
 static void UI_UpdateUserinfo(void);
