+++ xash3d_engine/engine/client/cl_render.c
@@ -17,6 +17,7 @@ GNU General Public License for more details.
 #include "client.h"
 #include "library.h"
 #include "platform/platform.h"
+#include "CRCLib/crclib.h"
 
 int R_FatPVS(const vec3_t org, float radius, byte* visbuffer, qboolean merge, qboolean fullvis)
 {
