+++ xash3d_engine/engine/client/vgui/vgui_draw.c
@@ -22,8 +22,8 @@ GNU General Public License for more details.
 #include "ref_common.h"
 #include "input.h"
 #include "platform/platform.h"
-#include "BuildDefs/libnames.h"
-#include "BuildDefs/build.h"
+#include "PlatformDefs/libnames.h"
+#include "PlatformDefs/build.h"
 
 #ifdef XASH_WIN32
 #define VGUI_SUPPORT_DLL "../vgui_support." OS_LIB_EXT
