+++ xash3d_engine/engine/common/filesystem_engine.c
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 
 #include "common.h"
 #include "library.h"
-#include "BuildDefs/libnames.h"
+#include "PlatformDefs/libnames.h"
 
 fs_api_t g_fsapi;
 fs_globals_t* FI;
