+++ xash3d_engine/engine/common/lib_common.c
@@ -19,7 +19,7 @@ GNU General Public License for more details.
 #include "server.h"
 #include <ctype.h>
 #include "CommonUtils/arch.h"
-#include "BuildDefs/libnames.h"
+#include "PlatformDefs/libnames.h"
 
 static char s_szLastError[1024] = "";
