+++ xash3d_engine/engine/common/pm_local.h
@@ -16,6 +16,7 @@ GNU General Public License for more details.
 #ifndef PM_LOCAL_H
 #define PM_LOCAL_H
 
+#include <string.h>
 #include "pm_defs.h"
 
 typedef int (*pfnIgnore)(physent_t* pe);  // custom trace filter
