+++ xash3d_engine/engine/menu_int.h
@@ -19,7 +19,7 @@ GNU General Public License for more details.
 #include "cvardef.h"
 #include "gameinfo.h"
 #include "wrect.h"
-#include "BuildDefs/decorators.h"
+#include "PlatformDefs/decorators.h"
 
 // a macro for mainui_cpp, indicating that mainui should be compiled for
 // Xash3D 1.0 interface
