+++ xash3d_engine/engine/ref/gl/CMakeLists.txt
@@ -50,6 +50,7 @@ target_link_libraries(${TARGETNAME_LIB_REF_OPENGL} PRIVATE
 	${TARGETNAME_INCLUDES_ENGINE}
 	${TARGETNAME_LIB_COMMONUTILS}
 	${TARGETNAME_INCLUDES_SDK}
+	${TARGETNAME_LIB_CRCLIB}
 
 	$<$<BOOL:${NF_OPENGL_STATIC_LINK}>:GL>
