+++ xash3d_engine/engine/ref/soft/CMakeLists.txt
@@ -42,6 +42,7 @@ target_link_libraries(${TARGETNAME_LIB_REF_SOFT} PRIVATE
 	${TARGETNAME_INCLUDES_ENGINE}
 	${TARGETNAME_LIB_COMMONUTILS}
 	${TARGETNAME_INCLUDES_SDK}
+	${TARGETNAME_LIB_CRCLIB}
 
 	# Linux
 	$<$<STREQUAL:"${TARGET_PLATFORM}","LINUX">:m>
