+++ xash3d_engine/engine/ref/soft/r_context.c
@@ -13,6 +13,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
+#include <string.h>
 #include "r_local.h"
 
 ref_api_t gEngfuncs;
