+++ xash3d_engine/engine/ref/soft/r_decals.c
@@ -13,6 +13,8 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
+#include <string.h>
+#include "CommonUtils/crtlib.h"
 #include "r_local.h"
 
 #define DECAL_OVERLAP_DISTANCE 2
