+++ xash3d_engine/engine/ref/soft/r_image.c
@@ -13,7 +13,10 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
+#include <string.h>
 #include "r_local.h"
+#include "CommonUtils/crtlib.h"
+#include "CRCLib/crclib.h"
 
 #define TEXTURES_HASH_SIZE (MAX_TEXTURES >> 2)
