+++ xash3d_engine/engine/ref/soft/r_misc.c
@@ -19,6 +19,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 // r_misc.c
 
+#include <string.h>
 #include "r_local.h"
 
 #define NUM_MIPS 4
