+++ xash3d_engine/engine/ref/soft/r_sprite.c
@@ -18,6 +18,7 @@ GNU General Public License for more details.
 #include "sprite.h"
 #include "studio.h"
 #include "entity_types.h"
+#include "CommonUtils/crtlib.h"
 
 // it's a Valve default value for LoadMapSprite (probably must be power of two)
 #define MAPSPRITE_SIZE 128
