+++ xash3d_engine/engine/ref/soft/r_studio.c
@@ -15,6 +15,7 @@ GNU General Public License for more details.
 
 #include "r_local.h"
 #include "CommonUtils/xash3d_mathlib.h"
+#include "CommonUtils/crtlib.h"
 #include "const.h"
 #include "r_studioint.h"
 #include "triangleapi.h"
