+++ xash3d_engine/engine/ref/soft/r_surf.c
@@ -21,6 +21,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "r_local.h"
 #include "mod_local.h"
+#include "CommonUtils/crtlib.h"
 
 drawsurf_t r_drawsurf;
