+++ xash3d_engine/engine/sprite.h
@@ -16,7 +16,7 @@
 #ifndef SPRITE_H
 #define SPRITE_H
 
-#include "BuildDefs/build.h"
+#include "PlatformDefs/build.h"
 #include <stdint.h>
 
 /*
