+++ xash3d_engine/filesystem/CMakeLists.txt
@@ -15,7 +15,7 @@ target_link_libraries(${TARGETNAME_LIB_FILESYSTEM_STDIO} PUBLIC
 	${TARGETNAME_INCLUDES_FILESYSTEM}
 	${TARGETNAME_INCLUDES_SDK}
 	${TARGETNAME_LIB_COMMONUTILS}
-	${TARGETNAME_LIB_BUILDDEFS}
+	${TARGETNAME_LIB_PLATFORMDEFS}
 	${TARGETNAME_LIB_CRCLIB}
 )
