+++ xash3d_engine/filesystem/pak.c
@@ -16,7 +16,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "BuildDefs/build.h"
+#include "PlatformDefs/build.h"
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
