+++ xash3d_engine/filesystem/zip.c
@@ -14,7 +14,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "BuildDefs/build.h"
+#include "PlatformDefs/build.h"
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
